Correctly implement Grid Sybil Self attack

+++ src/artery/application/md/F2MDParameters.h
@@ -70,7 +70,7 @@ namespace artery {
         double AttackGridSybilDistanceX;
         double AttackGridSybilDistanceY;
         double AttackGridSybilDistanceVariation;
-        int AttackGridSybilMaxDistanceFromRoad;
+        double AttackGridSybilMaxDistanceFromRoad;
     };
 
     struct DetectionParameters {
